+++ f4jumble_long.py
@@ -3,8 +3,8 @@ import sys; assert sys.version_info[0] >= 3, "Python 3 required."
 
 from pyblake2 import blake2b
 
-from tv_output import render_args, render_tv
-from f4jumble import f4jumble, f4jumble_inv, MAX_l_M
+from zcash_test_vectors.output import render_args, render_tv
+from zcash_test_vectors.f4jumble import f4jumble, f4jumble_inv, MAX_l_M
 
 
 def main():
