+++ orchard_empty_roots.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
 import sys; assert sys.version_info[0] >= 3, "Python 3 required."
 
-from orchard_merkle_tree import empty_roots
-from orchard_pallas import Fp
-from tv_output import render_args, render_tv
-from utils import i2lebsp
+from zcash_test_vectors.orchard.merkle_tree import empty_roots
+from zcash_test_vectors.orchard.pallas import Fp
+from zcash_test_vectors.output import render_args, render_tv
+from zcash_test_vectors.utils import i2lebsp
 
 
 def main():
