+++ pyproject.toml
@@ -27,3 +27,26 @@ numpy = "1.21.0"
 chacha20poly1305 = "0.0.3"
 cryptography = "36.0.0"
 pyblake2 = "1.1.2"
+
+[tool.poetry.scripts]
+# General test vectors
+f4jumble = "zcash_test_vectors.f4jumble:main"
+zip_0143 = "zcash_test_vectors.zip_0143:main"
+zip_0243 = "zcash_test_vectors.zip_0243:main"
+zip_0244 = "zcash_test_vectors.zip_0244:main"
+
+# Sapling test vectors
+sapling_generators = "zcash_test_vectors.sapling.generators:main"
+sapling_key_components = "zcash_test_vectors.sapling.key_components:main"
+sapling_note_encryption = "zcash_test_vectors.sapling.note_encryption:main"
+sapling_signatures = "zcash_test_vectors.sapling.redjubjub:main"
+sapling_zip32 = "zcash_test_vectors.sapling.zip32:main"
+
+# Orchard test vectors
+orchard_generators = "zcash_test_vectors.orchard.generators:main"
+orchard_group_hash = "zcash_test_vectors.orchard.group_hash:main"
+orchard_key_components = "zcash_test_vectors.orchard.key_components:main"
+orchard_merkle_tree = "zcash_test_vectors.orchard.merkle_tree:main"
+orchard_note_encryption = "zcash_test_vectors.orchard.note_encryption:main"
+orchard_poseidon = "zcash_test_vectors.orchard.poseidon:main"
+orchard_sinsemilla = "zcash_test_vectors.orchard.sinsemilla:main"
