+++ unified_address.py
@@ -5,16 +5,16 @@ import math
 from random import Random
 import struct
 
-from bech32m import bech32_encode, bech32_decode, convertbits, Encoding
+from zcash_test_vectors.bech32m import bech32_encode, bech32_decode, convertbits, Encoding
 
-from tv_output import render_args, render_tv, Some
-from tv_rand import Rand, randbytes
-from zc_utils import write_compact_size, parse_compact_size
-from f4jumble import f4jumble, f4jumble_inv
-import sapling_key_components
-import orchard_key_components
-from unified_encoding import encode_unified, decode_unified
-from unified_encoding import P2PKH_ITEM, P2SH_ITEM, SAPLING_ITEM, ORCHARD_ITEM
+from zcash_test_vectors.output import render_args, render_tv, Some
+from zcash_test_vectors.rand import Rand, randbytes
+from zcash_test_vectors.zc_utils import write_compact_size, parse_compact_size
+from zcash_test_vectors.f4jumble import f4jumble, f4jumble_inv
+from zcash_test_vectors.sapling import key_components as sapling_key_components
+from zcash_test_vectors.orchard import key_components as orchard_key_components
+from zcash_test_vectors.unified_encoding import encode_unified, decode_unified
+from zcash_test_vectors.unified_encoding import P2PKH_ITEM, P2SH_ITEM, SAPLING_ITEM, ORCHARD_ITEM
 
 def main():
     args = render_args()
