+++ zcash_test_vectors/bech32m.py
