+++ zcash_test_vectors/orchard/sinsemilla.py
@@ -3,13 +3,11 @@ import sys; assert sys.version_info[0] >= 3, "Python 3 required."
 
 import math
 
-import orchard_iso_pallas
-
-from orchard_pallas import Fp, Point
-from utils import cldiv, lebs2ip, i2leosp
-from orchard_group_hash import group_hash
-from tv_output import render_args, render_tv
-from tv_rand import Rand
+from .pallas import Fp, Point
+from ..utils import cldiv, lebs2ip, i2leosp
+from .group_hash import group_hash
+from ..output import render_args, render_tv
+from ..rand import Rand
 
 SINSEMILLA_K = 10
