+++ zcash_test_vectors/orchard/utils.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 import sys; assert sys.version_info[0] >= 3, "Python 3 required."
 
-from orchard_pallas import Fp, Scalar
-from utils import leos2ip
+from .pallas import Fp, Scalar
+from ..utils import leos2ip
 
 #
 # Utilities
