+++ zcash_test_vectors/sapling/key_components.py
@@ -3,12 +3,12 @@ import sys; assert sys.version_info[0] >= 3, "Python 3 required."
 
 from pyblake2 import blake2b, blake2s
 
-from sapling_generators import PROVING_KEY_BASE, SPENDING_KEY_BASE, group_hash
-from sapling_jubjub import Fr
-from sapling_merkle_tree import MERKLE_DEPTH
-from sapling_notes import note_commit, note_nullifier
-from utils import leos2bsp, leos2ip
-from tv_output import render_args, render_tv
+from .generators import PROVING_KEY_BASE, SPENDING_KEY_BASE, group_hash
+from .jubjub import Fr
+from .merkle_tree import MERKLE_DEPTH
+from .notes import note_commit, note_nullifier
+from ..utils import leos2bsp, leos2ip
+from ..output import render_args, render_tv
 
 #
 # Utilities
