+++ zcash_test_vectors/sapling/merkle_tree.py
@@ -3,8 +3,8 @@ import sys; assert sys.version_info[0] >= 3, "Python 3 required."
 
 from binascii import unhexlify
 
-from sapling_pedersen import pedersen_hash
-from utils import i2lebsp, leos2bsp
+from .pedersen import pedersen_hash
+from ..utils import i2lebsp, leos2bsp
 
 MERKLE_DEPTH = 32
