+++ zcash_test_vectors/zip_0243.py
@@ -4,16 +4,16 @@ import sys; assert sys.version_info[0] >= 3, "Python 3 required."
 from pyblake2 import blake2b
 import struct
 
-from transaction import (
+from .transaction import (
     LegacyTransaction,
     MAX_MONEY,
     SAPLING_TX_VERSION,
     Script,
 )
-from tv_output import render_args, render_tv, Some
-from tv_rand import Rand
+from .output import render_args, render_tv, Some
+from .rand import Rand
 
-from zip_0143 import (
+from .zip_0143 import (
     getHashJoinSplits,
     getHashOutputs,
     getHashPrevouts,
